refactor : move response io model to top level directory
