+++ src/apis/auth/mutations/useAuthJoinMutation.ts
@@ -2,7 +2,7 @@ import { useMutation } from "@tanstack/react-query";
 
 import { AuthJoinRequest } from "@apis/auth/models/AuthJoin";
 
-import { RestErrorResponse, RestResponse } from "@models/io/RestResponse";
+import { RestErrorResponse, RestResponse } from "@models/RestResponse";
 
 import rest from "@utils/rest";
