+++ src/apis/auth/queries/useAuthValidQuery.ts
@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 
 import { AuthValidResponse } from "@apis/auth/models/AuthValid";
 
-import { RestErrorResponse, RestResponse } from "@models/io/RestResponse";
+import { RestErrorResponse, RestResponse } from "@models/RestResponse";
 
 import rest from "@utils/rest";
